requirements.txt
streamlit
python-bcb
pandas
python-dateutil